refactor: Break up blob storage code into sub modules
