+++ tests/unit/test_storage_validation.py
@@ -1,30 +1,18 @@
-"""Tests for the storage module."""
+"""Security tests for the storage module."""
 
-import hashlib
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import MagicMock, patch
 
 import pytest
-from botocore.exceptions import ClientError
 
-from tracecat import storage
-from tracecat.storage import (
+from tracecat.storage.validation import (
     ALLOWED_CONTENT_TYPES,
+    ALLOWED_EXTENSIONS,
     BLOCKED_CONTENT_TYPES,
     BLOCKED_EXTENSIONS,
     MAX_FILE_SIZE,
     FileContentMismatchError,
     FileSecurityError,
     FileSecurityValidator,
-    compute_sha256,
-    delete_file,
-    download_file,
-    ensure_bucket_exists,
-    file_exists,
-    generate_presigned_download_url,
-    generate_presigned_upload_url,
-    upload_file,
-    validate_content_type,
-    validate_file_size,
 )
 
 
@@ -185,7 +173,7 @@ class TestFileSecurityValidator:
             result = self.validator._cross_validate_file_type(
                 extension, declared_type, detected_type
             )
-            expected_type = storage.ALLOWED_EXTENSIONS[extension]
+            expected_type = ALLOWED_EXTENSIONS[extension]
             assert result == expected_type
 
     def test_cross_validate_file_type_invalid(self):
@@ -358,294 +346,46 @@ class TestFileSecurityValidator:
                 declared_content_type="application/x-executable",
             )
 
+    def test_empty_file_content(self):
+        """Test handling of empty file content."""
+        with pytest.raises(ValueError, match="File cannot be empty"):
+            self.validator.validate_file(
+                content=b"", filename="empty.txt", declared_content_type="text/plain"
+            )
 
-class TestUtilityFunctions:
-    """Test utility functions."""
+    def test_unicode_filename(self):
+        """Test handling of Unicode filenames."""
+        # Fixed: \w includes Unicode word characters, so they're preserved
+        result = self.validator._sanitize_filename("测试文件.txt")
+        assert result == "测试文件.txt"  # Unicode chars are preserved by \w regex
 
-    def test_compute_sha256(self):
-        """Test SHA256 computation."""
-        content = b"hello world"
-        expected = hashlib.sha256(content).hexdigest()
-        result = compute_sha256(content)
-        assert result == expected
+    def test_very_long_filename(self):
+        """Test handling of very long filenames."""
+        long_name = "a" * 300 + ".txt"
+        result = self.validator._sanitize_filename(long_name)
+        assert len(result) <= 255
+        assert result.endswith(".txt")
 
-    def test_validate_content_type_valid(self):
-        """Test valid content type validation."""
-        for content_type in ALLOWED_CONTENT_TYPES:
-            validate_content_type(content_type)
+    def test_filename_sanitization_security(self):
+        """Test that filename sanitization removes security risks."""
+        # Test various security-problematic filenames
+        test_cases = [
+            # Control characters
+            ("file\x00name.txt", "filename.txt"),
+            ("file\x01\x02name.txt", "filename.txt"),
+            # Unicode normalization attacks
+            ("file\u202ename.txt", "filename.txt"),  # Right-to-left override
+            # Long filename attack
+            ("A" * 300 + ".txt", "A" * (255 - 4) + ".txt"),
+            # Hidden file attempts
+            ("...hidden.txt", "hidden.txt"),
+        ]
 
-        # With parameters
-        validate_content_type("text/plain; charset=utf-8")
-
-    def test_validate_content_type_invalid(self):
-        """Test invalid content type validation."""
-        # Blocked types
-        for blocked_type in BLOCKED_CONTENT_TYPES:
-            with pytest.raises(ValueError, match="not allowed for security reasons"):
-                validate_content_type(blocked_type)
-
-        # Unsupported type
-        with pytest.raises(ValueError, match="not supported"):
-            validate_content_type("application/unsupported")
-
-    def test_validate_file_size_function(self):
-        """Test standalone file size validation function."""
-        # Valid size
-        validate_file_size(1024)
-
-        # Too large
-        with pytest.raises(ValueError, match="exceeds maximum allowed size"):
-            validate_file_size(MAX_FILE_SIZE + 1)
-
-
-class TestS3Operations:
-    """Test S3/MinIO operations."""
-
-    @pytest.fixture
-    def mock_s3_client(self):
-        """Create a mock S3 client."""
-        mock_client = AsyncMock()
-        return mock_client
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_ensure_bucket_exists_existing(self, mock_get_client):
-        """Test ensure_bucket_exists when bucket already exists."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        await ensure_bucket_exists("test-bucket")
-        mock_client.head_bucket.assert_called_once_with(Bucket="test-bucket")
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_ensure_bucket_exists_create_new(self, mock_get_client):
-        """Test ensure_bucket_exists when bucket needs to be created."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        # Simulate bucket not found
-        mock_client.head_bucket.side_effect = ClientError(
-            error_response={"Error": {"Code": "404"}}, operation_name="head_bucket"
-        )
-
-        await ensure_bucket_exists("test-bucket")
-        mock_client.head_bucket.assert_called_once_with(Bucket="test-bucket")
-        mock_client.create_bucket.assert_called_once_with(Bucket="test-bucket")
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_upload_file(self, mock_get_client):
-        """Test file upload."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        content = b"test content"
-        key = "test/file.txt"
-        content_type = "text/plain"
-
-        await upload_file(content, key, "test-bucket", content_type)
-
-        mock_client.put_object.assert_called_once_with(
-            Bucket="test-bucket", Key=key, Body=content, ContentType=content_type
-        )
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_download_file(self, mock_get_client):
-        """Test file download."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        expected_content = b"test content"
-        mock_response = {"Body": AsyncMock()}
-        mock_response["Body"].read.return_value = expected_content
-        mock_client.get_object.return_value = mock_response
-
-        result = await download_file("test/file.txt", "test-bucket")
-
-        assert result == expected_content
-        mock_client.get_object.assert_called_once_with(
-            Bucket="test-bucket", Key="test/file.txt"
-        )
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_download_file_not_found(self, mock_get_client):
-        """Test file download when file doesn't exist."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        mock_client.get_object.side_effect = ClientError(
-            error_response={"Error": {"Code": "NoSuchKey"}}, operation_name="get_object"
-        )
-
-        with pytest.raises(FileNotFoundError, match="File not found"):
-            await download_file("nonexistent.txt", "test-bucket")
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_delete_file(self, mock_get_client):
-        """Test file deletion."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        await delete_file("test/file.txt", "test-bucket")
-
-        mock_client.delete_object.assert_called_once_with(
-            Bucket="test-bucket", Key="test/file.txt"
-        )
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_file_exists_true(self, mock_get_client):
-        """Test file_exists when file exists."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        result = await file_exists("test/file.txt", "test-bucket")
-
-        assert result is True
-        mock_client.head_object.assert_called_once_with(
-            Bucket="test-bucket", Key="test/file.txt"
-        )
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_file_exists_false(self, mock_get_client):
-        """Test file_exists when file doesn't exist."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        mock_client.head_object.side_effect = ClientError(
-            error_response={"Error": {"Code": "404"}}, operation_name="head_object"
-        )
-
-        result = await file_exists("nonexistent.txt", "test-bucket")
-
-        assert result is False
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    @patch("tracecat.storage.config")
-    async def test_generate_presigned_download_url(self, mock_config, mock_get_client):
-        """Test presigned download URL generation."""
-        # Mock config
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_EXPIRY = 10
-        mock_config.TRACECAT__PUBLIC_APP_URL = "http://localhost"
-        mock_config.TRACECAT__BLOB_STORAGE_ENDPOINT = "http://minio:9000"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_ENDPOINT = None
-
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        expected_url = "https://example.com/presigned-url"
-        mock_client.generate_presigned_url.return_value = expected_url
-
-        result = await generate_presigned_download_url(
-            "test/file.txt", "test-bucket", 3600
-        )
-
-        assert result == expected_url
-        mock_client.generate_presigned_url.assert_called_once_with(
-            "get_object",
-            Params={
-                "Bucket": "test-bucket",
-                "Key": "test/file.txt",
-                "ResponseContentDisposition": 'attachment; filename="file.txt"',
-            },
-            ExpiresIn=3600,
-        )
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    @patch("tracecat.storage.config")
-    async def test_generate_presigned_download_url_with_preview(
-        self, mock_config, mock_get_client
-    ):
-        """Test presigned download URL generation with preview mode."""
-        # Mock config
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_EXPIRY = 10
-        mock_config.TRACECAT__PUBLIC_APP_URL = "http://localhost"
-        mock_config.TRACECAT__BLOB_STORAGE_ENDPOINT = "http://minio:9000"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_ENDPOINT = None
-
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        expected_url = "https://example.com/presigned-url"
-        mock_client.generate_presigned_url.return_value = expected_url
-
-        # Test with preview mode disabled (force_download=False)
-        result = await generate_presigned_download_url(
-            "test/image.png", "test-bucket", 3600, force_download=False
-        )
-
-        assert result == expected_url
-        mock_client.generate_presigned_url.assert_called_once_with(
-            "get_object",
-            Params={
-                "Bucket": "test-bucket",
-                "Key": "test/image.png",
-                # No ResponseContentDisposition when force_download=False
-            },
-            ExpiresIn=3600,
-        )
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    @patch("tracecat.storage.config")
-    async def test_generate_presigned_download_url_with_content_type_override(
-        self, mock_config, mock_get_client
-    ):
-        """Test presigned download URL generation with content type override."""
-        # Mock config
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_EXPIRY = 10
-        mock_config.TRACECAT__PUBLIC_APP_URL = "http://localhost"
-        mock_config.TRACECAT__BLOB_STORAGE_ENDPOINT = "http://minio:9000"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_ENDPOINT = None
-
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        expected_url = "https://example.com/presigned-url"
-        mock_client.generate_presigned_url.return_value = expected_url
-
-        # Test with content type override
-        result = await generate_presigned_download_url(
-            "test/file.bin",
-            "test-bucket",
-            3600,
-            override_content_type="application/octet-stream",
-        )
-
-        assert result == expected_url
-        mock_client.generate_presigned_url.assert_called_once_with(
-            "get_object",
-            Params={
-                "Bucket": "test-bucket",
-                "Key": "test/file.bin",
-                "ResponseContentDisposition": 'attachment; filename="file.bin"',
-                "ResponseContentType": "application/octet-stream",
-            },
-            ExpiresIn=3600,
-        )
-
-    @patch("tracecat.storage.get_storage_client")
-    async def test_generate_presigned_upload_url(self, mock_get_client):
-        """Test presigned upload URL generation."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        expected_url = "https://example.com/presigned-upload-url"
-        mock_client.generate_presigned_url.return_value = expected_url
-
-        result = await generate_presigned_upload_url(
-            "test/file.txt", "test-bucket", 3600, "text/plain"
-        )
-
-        assert result == expected_url
-        mock_client.generate_presigned_url.assert_called_once_with(
-            "put_object",
-            Params={
-                "Bucket": "test-bucket",
-                "Key": "test/file.txt",
-                "ContentType": "text/plain",
-            },
-            ExpiresIn=3600,
-        )
+        for malicious_name, expected_safe in test_cases:
+            result = self.validator._sanitize_filename(malicious_name)
+            assert result == expected_safe
+            assert len(result) <= 255
+            assert ".." not in result
 
 
 class TestSecurityConfiguration:
@@ -716,123 +456,8 @@ class TestSecurityConfiguration:
         )
 
 
-class TestEdgeCases:
-    """Test edge cases and error conditions."""
-
-    def test_empty_file_content(self):
-        """Test handling of empty file content."""
-        validator = FileSecurityValidator()
-
-        with pytest.raises(ValueError, match="File cannot be empty"):
-            validator.validate_file(
-                content=b"", filename="empty.txt", declared_content_type="text/plain"
-            )
-
-    def test_unicode_filename(self):
-        """Test handling of Unicode filenames."""
-        validator = FileSecurityValidator()
-
-        # Fixed: \w includes Unicode word characters, so they're preserved
-        result = validator._sanitize_filename("测试文件.txt")
-        assert result == "测试文件.txt"  # Unicode chars are preserved by \w regex
-
-    def test_very_long_filename(self):
-        """Test handling of very long filenames."""
-        validator = FileSecurityValidator()
-
-        long_name = "a" * 300 + ".txt"
-        result = validator._sanitize_filename(long_name)
-        assert len(result) <= 255
-        assert result.endswith(".txt")
-
-    @patch("tracecat.storage.logger")
-    def test_logging_on_errors(self, mock_logger):
-        """Test that errors are properly logged."""
-        with patch("tracecat.storage.get_storage_client") as mock_get_client:
-            mock_client = AsyncMock()
-            mock_get_client.return_value.__aenter__.return_value = mock_client
-
-            # Simulate S3 error
-            mock_client.put_object.side_effect = ClientError(
-                error_response={"Error": {"Code": "AccessDenied"}},
-                operation_name="put_object",
-            )
-
-            # This should raise and log
-            with pytest.raises(ClientError):
-                import asyncio
-
-                asyncio.run(upload_file(b"content", "key", "test"))
-
-            # Check that error was logged
-            mock_logger.error.assert_called()
-
-
 class TestSecurityHardening:
     """Test security hardening features for blob storage."""
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    @patch("tracecat.storage.config")
-    async def test_presigned_url_path_replacement(self, mock_config, mock_get_client):
-        """Test that presigned URLs correctly include /s3 path prefix."""
-        # Setup configuration
-        mock_config.TRACECAT__PUBLIC_APP_URL = "http://localhost"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_EXPIRY = 10
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_ENDPOINT = (
-            "http://localhost/s3"
-        )
-        mock_config.TRACECAT__BLOB_STORAGE_PROTOCOL = "minio"
-        mock_config.TRACECAT__BLOB_STORAGE_ENDPOINT = "http://minio:9000"
-
-        # Setup mock client
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        # Mock the original presigned URL from MinIO
-        original_url = "http://minio:9000/tracecat/attachments/test.txt?AWSAccessKeyId=minio&Signature=abc123&Expires=1234567890"
-        mock_client.generate_presigned_url.return_value = original_url
-
-        # Generate presigned URL
-        result = await generate_presigned_download_url(
-            "attachments/test.txt", "tracecat", 30
-        )
-
-        # Verify URL transformation
-        expected_url = "http://localhost/s3/tracecat/attachments/test.txt?AWSAccessKeyId=minio&Signature=abc123&Expires=1234567890"
-        assert result == expected_url
-        assert "/s3/" in result
-        assert "localhost" in result
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    @patch("tracecat.storage.config")
-    async def test_presigned_url_no_replacement_when_not_minio(
-        self, mock_config, mock_get_client
-    ):
-        """Test that URLs are not modified when not from internal MinIO."""
-        # Mock config
-        mock_config.TRACECAT__PUBLIC_APP_URL = "http://localhost"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_EXPIRY = 10
-        mock_config.TRACECAT__BLOB_STORAGE_ENDPOINT = "http://minio:9000"
-        mock_config.TRACECAT__BLOB_STORAGE_PRESIGNED_URL_ENDPOINT = None
-
-        # Setup mock client
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        # URL that doesn't start with http://minio:9000
-        original_url = "https://s3.amazonaws.com/bucket/test.txt?AWSAccessKeyId=key&Signature=abc123&Expires=1234567890"
-        mock_client.generate_presigned_url.return_value = original_url
-
-        # Generate presigned URL
-        result = await generate_presigned_download_url(
-            "attachments/test.txt", "tracecat", 30
-        )
-
-        # URL should remain unchanged since it's not from MinIO
-        assert result == original_url
-        assert "s3.amazonaws.com" in result
 
     def test_file_validation_against_malicious_content(self):
         """Test comprehensive file validation against various attack vectors."""
@@ -896,28 +521,6 @@ class TestSecurityHardening:
                     declared_content_type="text/plain",
                 )
 
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    async def test_storage_error_handling_security(self, mock_get_client):
-        """Test that storage errors don't leak sensitive information."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        # Simulate various S3 errors
-        mock_client.get_object.side_effect = ClientError(
-            error_response={
-                "Error": {
-                    "Code": "AccessDenied",
-                    "Message": "Access denied to sensitive-bucket/secret-path/",
-                }
-            },
-            operation_name="get_object",
-        )
-
-        # Error should be raised but internal paths should not be exposed
-        with pytest.raises(ClientError):
-            await download_file("test-file.txt", "test-bucket")
-
     def test_file_size_limits_enforcement(self):
         """Test that file size limits are strictly enforced."""
         validator = FileSecurityValidator()
@@ -945,52 +548,3 @@ class TestSecurityHardening:
                 filename="archive.zip",
                 declared_content_type="application/zip",
             )
-
-    @pytest.mark.anyio
-    @patch("tracecat.storage.get_storage_client")
-    async def test_presigned_url_expiry_validation(self, mock_get_client):
-        """Test that presigned URLs have appropriate expiry times."""
-        mock_client = AsyncMock()
-        mock_get_client.return_value.__aenter__.return_value = mock_client
-
-        mock_client.generate_presigned_url.return_value = (
-            "http://test.com/file?Expires=123"
-        )
-
-        # Test with various expiry times
-        short_expiry = 30  # 30 seconds - appropriate for immediate download
-        await generate_presigned_download_url("test.txt", "bucket", short_expiry)
-
-        # Verify the expiry was passed correctly with security headers
-        mock_client.generate_presigned_url.assert_called_with(
-            "get_object",
-            Params={
-                "Bucket": "bucket",
-                "Key": "test.txt",
-                "ResponseContentDisposition": 'attachment; filename="test.txt"',
-            },
-            ExpiresIn=short_expiry,
-        )
-
-    def test_filename_sanitization_security(self):
-        """Test that filename sanitization removes security risks."""
-        validator = FileSecurityValidator()
-
-        # Test various security-problematic filenames
-        test_cases = [
-            # Control characters
-            ("file\x00name.txt", "filename.txt"),
-            ("file\x01\x02name.txt", "filename.txt"),
-            # Unicode normalization attacks
-            ("file\u202ename.txt", "filename.txt"),  # Right-to-left override
-            # Long filename attack
-            ("A" * 300 + ".txt", "A" * (255 - 4) + ".txt"),
-            # Hidden file attempts
-            ("...hidden.txt", "hidden.txt"),
-        ]
-
-        for malicious_name, expected_safe in test_cases:
-            result = validator._sanitize_filename(malicious_name)
-            assert result == expected_safe
-            assert len(result) <= 255
-            assert ".." not in result
